+++ models/Location.ts
@@ -1,21 +1,14 @@
 export type Location = {
   latitude: number;
   longitude: number;
-  city: string;
-  postalCode: string;
-  street: string;
-  streetNumber: string;
+  city?: string | null;
+  postalCode?: string | null;
+  street?: string | null;
+  streetNumber?: string | null;
 };
 
-export const defaultLocation: Location = {
-  latitude: -1,
-  longitude: -1,
-  city: '',
-  postalCode: '',
-  street: '',
-  streetNumber: '',
-};
-
-export const getAddressDisplayText = (location: Location): string => {
-  return `${location.street} ${location.streetNumber} ${location.city} ${location.postalCode}`.trim();
+export const getFormattedAddress = (location: Location): string => {
+  return `${location.street} ${location.streetNumber}, ${location.postalCode} ${location.city}`
+    .replace('  ', ' ')
+    .trim();
 };
