+++ navigation/index.tsx
@@ -3,7 +3,7 @@ import { Modal } from 'react-native';
 import { DefaultTheme, NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { AuthenticationScreen } from '../screens/AuthenticationScreen';
-import { MainScreen } from '../screens/MainScreen';
+import { MapScreen } from '../screens/MapScreen';
 import { NotFoundScreen } from '../screens/NotFoundScreen';
 import { RootStackParamList } from '../types';
 import LinkingConfiguration from './LinkingConfiguration';
@@ -22,30 +22,28 @@ export default function Navigation() {
  */
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
-function RootNavigator() {
-  return (
-    <Stack.Navigator>
-      <Stack.Screen
-        name="Authentication"
-        component={AuthenticationScreen}
-        options={{ headerShown: false }}
-      />
+const RootNavigator = () => (
+  <Stack.Navigator>
+    <Stack.Screen
+      name="Authentication"
+      component={AuthenticationScreen}
+      options={{ headerShown: false }}
+    />
 
-      <Stack.Screen
-        name="NotFound"
-        component={NotFoundScreen}
-        options={{ title: 'Oops!' }}
-      />
+    <Stack.Screen
+      name="NotFound"
+      component={NotFoundScreen}
+      options={{ title: 'Oops!' }}
+    />
 
-      <Stack.Screen
-        name="Main"
-        component={MainScreen}
-        options={{ headerShown: false }}
-      />
+    <Stack.Screen
+      name="Main"
+      component={MapScreen}
+      options={{ headerShown: false }}
+    />
 
-      <Stack.Group screenOptions={{ presentation: 'modal' }}>
-        <Stack.Screen name="Modal" component={Modal} />
-      </Stack.Group>
-    </Stack.Navigator>
-  );
-}
+    <Stack.Group screenOptions={{ presentation: 'modal' }}>
+      <Stack.Screen name="Modal" component={Modal} />
+    </Stack.Group>
+  </Stack.Navigator>
+);
